backend/Grocket/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
Babel==2.12.1
billiard==3.6.4.0
bleach==5.0.1
cached-property==1.5.2
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-bleach==3.0.1
django-ckeditor==6.5.1
django-countries==7.5.1
django-dump-load-utf8==0.0.4
django-filter==22.1
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-money==3.0.0
django-mptt==0.14.0
django-mptt-admin==2.4.1
django-phonenumber-field==7.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.3.0
drf-extra-fields==3.4.1
eventlet==0.33.3
flake8==5.0.4
greenlet==2.0.2
gunicorn==20.1.0
id==1.0.0
idna==3.4
importlib-metadata==6.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==1.1.1
mypy-extensions==1.0.0
oauthlib==3.2.2
phonenumbers==8.13.7
Pillow==9.4.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
py-money==0.5.0
py-moneyed==2.0
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.6
pyflakes==2.5.0
PyJWT==2.6.0
python-dotenv==1.0.0
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sqlparse==0.4.3
tinycss2==1.1.1
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
zipp==3.15.0